requirements_dev.txt
-rrequirements.txt
tox==2.7.0
yamllint==1.6.1
Sphinx==1.5.5

# Python 2.6 Support
astroid==1.2.1; python_version < '2.7' # pyup: <=1.2.1
pylint==1.3.1; python_version < '2.7' # pyup: >=1.3.0,<1.4.0
flake8==2.6.2; python_version < '2.7' # pyup: >=2.6.2,<3.0.0

# Python 2.7+ Support
pylint==1.7.1; python_version >= '2.7'
flake8==3.3.0; python_version >= '2.7'
